# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MedicAid",
        "description": "Documentación de la API de MedicAid (Usuarios, Medicamentos, Tratamientos, Historiales y Auditoría)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/medicamentos": {
            "get": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Listar medicamentos",
                "operationId": "c64f62e0d1a3b6f17e40b3ae98733ab2",
                "responses": {
                    "200": {
                        "description": "Listado de medicamentos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Crear medicamento",
                "operationId": "15e861c0d4becaccba7019d18f082289",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Paracetamol 500mg"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Analgesico y antipirético"
                                    },
                                    "via_administracion": {
                                        "type": "string",
                                        "example": "Oral"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Medicamento creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/medicamentos/{id}": {
            "get": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Obtener un medicamento",
                "operationId": "e1c8c6e6a69148615e417d2f2b36e560",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicamento encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Actualizar medicamento",
                "operationId": "8702d56a76d4a1ec6e195692982adbec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "via_administracion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Medicamento actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Medicamentos"
                ],
                "summary": "Eliminar medicamento",
                "operationId": "d05f9f3f4c4fcf138fd5fef29c4c0506",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tratamientos": {
            "get": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Listar tratamientos",
                "operationId": "7891c37499a819c50229db25fe8109fc",
                "responses": {
                    "200": {
                        "description": "Listado paginado de tratamientos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Crear tratamiento",
                "operationId": "fa31dcb091355a40afd3130f73e7fa4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "medicamentos"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medicamentos": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "dosis": {
                                                    "type": "string",
                                                    "example": "1 tableta cada 8 horas"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Tratamiento para infección respiratoria"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tratamiento creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tratamientos/{id}": {
            "get": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener un tratamiento",
                "operationId": "3385ad30ccacdc603687449aa7d68e6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tratamiento encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Actualizar tratamiento",
                "operationId": "6a439236d7742e294bb15a1d6df922a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tratamiento actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Eliminar tratamiento",
                "operationId": "74fc72f1a89faff821b39f0af3e2d4ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}